pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
Markdown = "^3.4.1"
todo = "^0.1"
[tool.poetry.dev-dependencies]
